requirements.txt
seaborn
tensorflow-cpu
scikit-learn
numpy
pandas
matplotlib
plotly